Moved graph view to components
