+++ src/app/modules/fsa2/FSAModule.js
@@ -44,7 +44,7 @@ import FSANodeInputHandler from './graph/inputs/FSANodeInputHandler.js';
 import FSAInitialMarkerLayer from './graph/renderer/FSAInitialMarkerLayer.js';
 import FSALabelEditorRenderer from './renderer/FSALabelEditorRenderer.js';
 
-import GraphView from 'graph2/components/views/GraphView.js';
+import GraphView from 'graph2/components/GraphView.js';
 import GraphNodeLayer from 'graph2/components/layers/GraphNodeLayer.js';
 import GraphEdgeLayer from 'graph2/components/layers/GraphEdgeLayer.js';
 import SelectionBoxLayer from 'graph2/components/layers/SelectionBoxLayer.js';
